LC-largest-odd-number-in-string [Runtime: N/A; Memory: N/A]

+++ greedy/largest-odd-number-in-string.cpp
@@ -2,12 +2,12 @@ class Solution {
 public:
     string largestOddNumber(string num) {
         string ans = "";
-        int temp = 0;
+        long long temp = 0;
         for (int i = 0; i < num.size(); i++) {
-            int n = num[i] - '0';
+            long long n = num[i] - '0';
             temp = (temp * 10) + n;
         }
-        int maxnum = INT_MIN;
+        long long maxnum = INT_MIN;
         while (temp > 0) {
             if (temp % 2 != 0) {
                 maxnum = max(maxnum, temp);
